Фикс форматирования qty/quantity/order_num при отправке и разборе транзакций QUIK
- ShmQuikTransport реально использует QuikJson.Options (раньше конструктор его игнорировал и строил свой JsonSerializerOptions без кастомных конвертеров)
- StringToIntConverter терпимее к дробно-отформатированным числам ("qty":1.0), StringToLongConverter — новый, для order_num/trade_num
- Transaction.QUANTITY сериализуется строкой (как PRICE) — иначе QUIK получает нестроковое значение и отклоняет заявку ("Неправильно указано количество")
- Order/Trade/TransactionReply/NegTrade: OrderNum/TradeNum/FirstOrderNum из double в long — double теряет точность на 18-19-значных номерах заявок, из-за чего KillOrder уходил с неверным ORDER_KEY и отмена молча проваливалась

+++ DataStructures/Transaction/NegTrade.cs
@@ -268,7 +268,7 @@ namespace QuikSharp.DataStructures.Transaction
         /// Номер заявки в торговой системе
         /// </summary>
         [JsonPropertyName("order_num")]
-        public double OrderNum { get; set; }
+        public long OrderNum { get; set; }
 
         /// <summary>
         /// Дата заключения
