+++ DataStructures/Transaction/Trade.cs
@@ -17,13 +17,13 @@ namespace QuikSharp.DataStructures.Transaction
         /// Номер сделки в торговой системе
         /// </summary>
         [JsonPropertyName("trade_num")]
-        public double TradeNum { get; set; }
+        public long TradeNum { get; set; }
 
         /// <summary>
         /// Номер заявки в торговой системе
         /// </summary>
         [JsonPropertyName("order_num")]
-        public double OrderNum { get; set; }
+        public long OrderNum { get; set; }
 
         /// <summary>
         /// Поручение/комментарий, обычно: код клиента/номер поручения
